+++ spec/ext/feature_branch__feature_branch_per_user_spec.rb
@@ -36,13 +36,13 @@ describe 'feature_branch object extensions' do
       it 'feature branches correctly after storing feature configuration per user in a separate process (ensuring persistence)' do
         AbstractFeatureBranch.configuration.feature_store_live_fetching = false
         AbstractFeatureBranch.load_application_features
-        scope_id = 1 # test the case of numeric ID
+        scope = 1 # test the case of numeric ID
         ruby_code = <<-RUBY_CODE
           $:.unshift('.')
           require 'redis'
           require 'lib/abstract_feature_branch'
           AbstractFeatureBranch.feature_store = Redis.new
-          AbstractFeatureBranch.toggle_features_for_scope('#{scope_id}',
+          AbstractFeatureBranch.toggle_features_for_scope('#{scope}',
                                                           :feature1 => false,
                                                           :feature3 => true,
                                                           :feature6 => true,
@@ -51,14 +51,20 @@ describe 'feature_branch object extensions' do
         RUBY_CODE
         system "ruby -e \"#{ruby_code}\""
         
+        AbstractFeatureBranch.toggled_features_for_scope(scope).should == ['feature3', 'feature6']
+        AbstractFeatureBranch.toggled_features_for_scope('unusedscope').should == []
+        AbstractFeatureBranch.scopes_for_feature('feature3').should == [scope.to_s]
+        AbstractFeatureBranch.scopes_for_feature('feature6').should == [scope.to_s]
+        AbstractFeatureBranch.scopes_for_feature('unusedscope').should == []
+
         features_enabled = []
-        feature_branch :feature1, scope_id do
+        feature_branch :feature1, scope do
           features_enabled << :feature1
         end
-        feature_branch :feature3, scope_id do
+        feature_branch :feature3, scope do
           features_enabled << :feature3
         end
-        feature_branch :feature6, scope_id do
+        feature_branch :feature6, scope do
           features_enabled << :feature6
         end
         feature_branch :feature6, 'otheruser@example.com' do
@@ -67,7 +73,7 @@ describe 'feature_branch object extensions' do
         feature_branch :feature6 do
           features_enabled << :feature6_nouserspecified
         end
-        feature_branch :feature7, scope_id do
+        feature_branch :feature7, scope do
           features_enabled << :feature7
         end
         features_enabled.should include(:feature1) #remains like features.yml
@@ -81,13 +87,13 @@ describe 'feature_branch object extensions' do
         AbstractFeatureBranch.load_application_features
         
         features_enabled = []
-        feature_branch :feature1, scope_id do
+        feature_branch :feature1, scope do
           features_enabled << :feature1
         end
-        feature_branch :feature3, scope_id do
+        feature_branch :feature3, scope do
           features_enabled << :feature3
         end
-        feature_branch :feature6, scope_id do
+        feature_branch :feature6, scope do
           features_enabled << :feature6
         end
         feature_branch :feature6, 'otheruser@example.com' do
@@ -96,7 +102,7 @@ describe 'feature_branch object extensions' do
         feature_branch :feature6 do
           features_enabled << :feature6_nouserspecified
         end
-        feature_branch :feature7, scope_id do
+        feature_branch :feature7, scope do
           features_enabled << :feature7
         end
         features_enabled.should include(:feature1) #remains like features.yml
@@ -109,28 +115,34 @@ describe 'feature_branch object extensions' do
       it 'update feature branching (disabling some features) after having stored feature configuration per user in a separate process (ensuring persistence)' do
         AbstractFeatureBranch.configuration.feature_store_live_fetching = false
         AbstractFeatureBranch.load_application_features
-        scope_id = 1 # test the case of numeric ID
+        scope = 1 # test the case of numeric ID
         ruby_code = <<-RUBY_CODE
           $:.unshift('.')
           require 'redis'
           require 'lib/abstract_feature_branch'
           AbstractFeatureBranch.feature_store = Redis.new
-          AbstractFeatureBranch.toggle_features_for_scope('#{scope_id}',
+          AbstractFeatureBranch.toggle_features_for_scope('#{scope}',
                                                           :feature6 => true,
                                                           :feature7 => false
                                                          )
-          AbstractFeatureBranch.toggle_features_for_scope('#{scope_id}',
+          AbstractFeatureBranch.toggle_features_for_scope('#{scope}',
                                                           :feature6 => false,
                                                           :feature7 => true
                                                          )
         RUBY_CODE
         system "ruby -e \"#{ruby_code}\""
         
+        AbstractFeatureBranch.toggled_features_for_scope(scope).should == ['feature7']
+        AbstractFeatureBranch.toggled_features_for_scope('unusedscope').should == []
+        AbstractFeatureBranch.scopes_for_feature('feature6').should == []
+        AbstractFeatureBranch.scopes_for_feature('feature7').should == [scope.to_s]
+        AbstractFeatureBranch.scopes_for_feature('unusedscope').should == []
+        
         features_enabled = []
-        feature_branch :feature6, scope_id do
+        feature_branch :feature6, scope do
           features_enabled << :feature6
         end
-        feature_branch :feature7, scope_id do
+        feature_branch :feature7, scope do
           features_enabled << :feature7
         end
         features_enabled.should_not include(:feature6) # remains like features.yml
@@ -140,10 +152,10 @@ describe 'feature_branch object extensions' do
         AbstractFeatureBranch.load_application_features
 
         features_enabled = []
-        feature_branch :feature6, scope_id do
+        feature_branch :feature6, scope do
           features_enabled << :feature6
         end
-        feature_branch :feature7, scope_id do
+        feature_branch :feature7, scope do
           features_enabled << :feature7
         end
         features_enabled.should_not include(:feature6)
@@ -154,13 +166,13 @@ describe 'feature_branch object extensions' do
     context 'per user with feature_store_live_fetching true' do
       it 'feature branches correctly after storing feature configuration per user in a separate process (ensuring persistence)' do
         AbstractFeatureBranch.configuration.feature_store_live_fetching = true
-        scope_id = 1 # test the case of numeric ID
+        scope = 1 # test the case of numeric ID
         ruby_code = <<-RUBY_CODE
           $:.unshift('.')
           require 'redis'
           require 'lib/abstract_feature_branch'
           AbstractFeatureBranch.feature_store = Redis.new
-          AbstractFeatureBranch.toggle_features_for_scope('#{scope_id}',
+          AbstractFeatureBranch.toggle_features_for_scope('#{scope}',
                                                           :feature1 => false,
                                                           :feature3 => true,
                                                           :feature6 => true,
@@ -169,14 +181,20 @@ describe 'feature_branch object extensions' do
         RUBY_CODE
         system "ruby -e \"#{ruby_code}\""
         
+        AbstractFeatureBranch.toggled_features_for_scope(scope).should == ['feature3', 'feature6']
+        AbstractFeatureBranch.toggled_features_for_scope('unusedscope').should == []
+        AbstractFeatureBranch.scopes_for_feature('feature3').should == [scope.to_s]
+        AbstractFeatureBranch.scopes_for_feature('feature6').should == [scope.to_s]
+        AbstractFeatureBranch.scopes_for_feature('unusedscope').should == []
+        
         features_enabled = []
-        feature_branch :feature1, scope_id do
+        feature_branch :feature1, scope do
           features_enabled << :feature1
         end
-        feature_branch :feature3, scope_id do
+        feature_branch :feature3, scope do
           features_enabled << :feature3
         end
-        feature_branch :feature6, scope_id do
+        feature_branch :feature6, scope do
           features_enabled << :feature6
         end
         feature_branch :feature6, 'otheruser@example.com' do
@@ -185,7 +203,7 @@ describe 'feature_branch object extensions' do
         feature_branch :feature6 do
           features_enabled << :feature6_nouserspecified
         end
-        feature_branch :feature7, scope_id do
+        feature_branch :feature7, scope do
           features_enabled << :feature7
         end
         features_enabled.should include(:feature1) #remains like features.yml
@@ -198,28 +216,34 @@ describe 'feature_branch object extensions' do
       
       it 'update feature branching (disabling some features) after having stored feature configuration per user in a separate process (ensuring persistence)' do
         AbstractFeatureBranch.configuration.feature_store_live_fetching = true
-        scope_id = 1 # test the case of numeric ID
+        scope = 1 # test the case of numeric ID
         ruby_code = <<-RUBY_CODE
           $:.unshift('.')
           require 'redis'
           require 'lib/abstract_feature_branch'
           AbstractFeatureBranch.feature_store = Redis.new
-          AbstractFeatureBranch.toggle_features_for_scope('#{scope_id}',
+          AbstractFeatureBranch.toggle_features_for_scope('#{scope}',
                                                           :feature6 => true,
                                                           :feature7 => false
                                                          )
-          AbstractFeatureBranch.toggle_features_for_scope('#{scope_id}',
+          AbstractFeatureBranch.toggle_features_for_scope('#{scope}',
                                                           :feature6 => false,
                                                           :feature7 => true
                                                          )
         RUBY_CODE
         system "ruby -e \"#{ruby_code}\""
         
+        AbstractFeatureBranch.toggled_features_for_scope(scope).should == ['feature7']
+        AbstractFeatureBranch.toggled_features_for_scope('unusedscope').should == []
+        AbstractFeatureBranch.scopes_for_feature('feature6').should == []
+        AbstractFeatureBranch.scopes_for_feature('feature7').should == [scope.to_s]
+        AbstractFeatureBranch.scopes_for_feature('unusedscope').should == []
+        
         features_enabled = []
-        feature_branch :feature6, scope_id do
+        feature_branch :feature6, scope do
           features_enabled << :feature6
         end
-        feature_branch :feature7, scope_id do
+        feature_branch :feature7, scope do
           features_enabled << :feature7
         end
         features_enabled.should_not include(:feature6)